LibGPU+LibSoftGPU: Move LightModelParameters into LibGPU
